+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 3.2 (upcoming)
+* Added: New option ( `--timeout`) for setting a number of seconds to wait before timeout connection (default 30).
 * Revised: Increased default timeout to 30 seconds.
 * Fixed: Bug-fix regarding Basic HTTP authentication.
 * Fixed: Bug-fix regarding connection problems.